+++ src/debug.ts
@@ -1,8 +1,14 @@
 import { Keyboard } from "isaac-typescript-definitions";
-import { printConsole } from "isaacscript-common";
+import {
+  log,
+  printConsole,
+  saveDataManagerSetGlobal,
+  setLogFunctionsGlobal,
+} from "isaacscript-common";
+import g from "./globals";
 import { hotkeys } from "./modConfigMenu";
 
-/** Currently, F2 is set up to execute this function. */
+/** Currently, F2 is set to execute this function. */
 export function debugCode(_params?: string): void {
   // Add code here.
 }
@@ -18,4 +24,14 @@ export function hotkey2Function(): void {
   hotkeys.roll = Keyboard.G;
 
   printConsole("Test hotkeys set.");
+}
+
+export function debugFunction(params?: string): void {
+  g.debug = true;
+  saveDataManagerSetGlobal();
+  setLogFunctionsGlobal();
+
+  log("Entering debug function.");
+  debugCode(params);
+  log("Exiting debug function.");
 }