+++ src/features/mandatory/extraConsoleCommands.ts
@@ -19,7 +19,7 @@ import {
 } from "isaacscript-common";
 import { logRaceData, RaceData } from "../../classes/RaceData";
 import { VERSION } from "../../constants";
-import { debugFunction } from "../../debugFunction";
+import { debugFunction } from "../../debug";
 import { ChallengeCustom } from "../../enums/ChallengeCustom";
 import { RaceFormat } from "../../enums/RaceFormat";
 import { RaceGoal } from "../../enums/RaceGoal";